♻️ Extract methods (#1346)
